Rename sourcetext.rb to source_text.rb
