+++ lib/racc/source_text.rb
